PHPstan level 7

+++ src/Console/CacheClearCommand.php
@@ -13,7 +13,8 @@ class CacheClearCommand extends Command
 {
     public function __construct(
         private readonly Settings $settings
-    ) {
+    )
+    {
         parent::__construct();
     }
 
@@ -33,7 +34,11 @@ class CacheClearCommand extends Command
 
     private function removeDirectory(string $path): void
     {
-        $files = glob($path.'/*');
+        $files = glob($path . '/*');
+        if ($files === false) {
+            return;
+        }
+
         foreach ($files as $file) {
             is_dir($file) ? $this->removeDirectory($file) : unlink($file);
         }
